Try to make it run in Python 2
